pyproject.toml
[project]
name = "notpatrick"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4",
    "ffmpeg-python>=0.2.0",
    "lxml>=6.0.0",
    "openai>=1.93.3",
    "openai-whisper",
    "requests",
    "chonkie[semantic]",
    "numpy",
    "pinecone",
    "pinecone-plugin-assistant",
    "python-dotenv",
    "cohere",
    "llama-index-core",
    "llama-index-llms-openai",
    "llama-index-vector-stores-pinecone",
    "llama-index-postprocessor-cohere-rerank",
    "llama-index-embeddings-openai>=0.3.1",
    "ruff>=0.12.2",
    "fastmcp",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
